Fix: module render post
